adding testUtilities.go.
